Removed current directory from all path alternatives
Removed the current directory options for GFX, SFX etc, to prevent
exploits by feeding the game bad data. (This was only intended as a
convenience for running without proper install while developing.)

This is an adaption of a Debian security patch. (I just removed the
code instead of commenting it out.)