Moved from Autotools to CMake
Initial move from Autotools to CMake. It Builds Here(TM) for Linux and
Windows (via MXE), and the binaries seem to mostly work, but there are
issues with paths, package building is most likely completely broken,
etc, etc. Some of these issues will be ignored until after the SDL2
port, as there are solutions in Kobo Redux that depend on SDL2.

  * Added "no longer maintained" notes to ChangeLog and TODO
  * Removed Autotools build system
  * Removed some obsolete or unused files
  * Added CMake build system, scripts, icons etc

Closes #4.